+++ src/sys/os/term/grid/_tests.rs
@@ -1,4 +1,4 @@
-// devela/src/sys/os/term/grid/tests.rs
+// devela/src/sys/os/term/grid/_tests.rs
 
 use crate::{TermColor, TermColors, TermGrid, TermGridError, TermPen, TermStyle, Termel};
 use crate::{ext, pos, region};